feat: add perfectionist/sort-named-imports and sort-named-exports
https://perfectionist.dev/rules/sort-named-imports
https://perfectionist.dev/rules/sort-named-exports

We had named imports sorted in v1 when tslint was taking care of that. But it got lost during migration to eslint, as I wrongly assumed that import/order will take care of that.
Later on, the support for sorting named imports (but still not exports) was added to the original eslint-plugin-import https://github.com/import-js/eslint-plugin-import/pull/3043
but we have switched to eslint-plugin-import-x since and the porting request of that feature is still pending. https://github.com/un-ts/eslint-plugin-import-x/issues/225

We could consider switching back to eslint-plugin-import but it still has a set of its own, still unresolved issues.
Instead, I decided to try eslint-plugin-perfectionist and the sorting rules from that. Note we are sticking with import/order for sorting and grouping imports for the time being, as more time would be needed to convert the sorting and grouping config to achieve similar behaviour with the perfectionist/sort-imports rule.

There is also perfectionist/sort-exports rule, but it does not provide config options to define grouping and custom sorting, to achieve similar behaviour to import/order with our custom config applied, so for now we won't enforce that rule.